# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "HOTEL",
        "description": "API for hotel management",
        "version": "1.0"
    },
    "produces": [
        "application/json"
    ],
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/clients": {
            "get": {
                "operationId": "allClients",
                "tags": [
                    "Clients"
                ],
                "description": "Recupere la liste des clients",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id_client": {
                                        "type": "integer"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "registration_date": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "operationId": "addClient",
                "tags": [
                    "Clients"
                ],
                "description": "Ajoute un client",
                "parameters": [
                    {
                        "name": "client",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "first_name": {
                                    "type": "string"
                                },
                                "last_name": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "phone": {
                                    "type": "string"
                                },
                                "registration_date": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/clients/{id}": {
            "get": {
                "operationId": "clientsById",
                "tags": [
                    "Clients"
                ],
                "description": "Recupere la liste des clients par leur id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id_client": {
                                        "type": "integer"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "registration_date": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "patch": {
                "operationId": "updateClient",
                "tags": [
                    "Clients"
                ],
                "description": "Met a jour un client",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "client",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "first_name": {
                                    "type": "string"
                                },
                                "last_name": {
                                    "type": "string"
                                },
                                "email": {
                                    "type": "string"
                                },
                                "phone": {
                                    "type": "string"
                                },
                                "registration_date": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "operationId": "deleteClient",
                "tags": [
                    "Clients"
                ],
                "description": "Supprime un client",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/clients/year/{year}": {
            "get": {
                "operationId": "clientsByYear",
                "tags": [
                    "Clients"
                ],
                "description": "Recupere la liste des clients inscrits une annee donnee",
                "parameters": [
                    {
                        "name": "year",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id_client": {
                                        "type": "integer"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "registration_date": {
                                        "type": "string"
                                    },
                                    "id_reservation": {
                                        "type": "integer"
                                    },
                                    "id_room": {
                                        "type": "integer"
                                    },
                                    "checkin_date": {
                                        "type": "string"
                                    },
                                    "checkout_date": {
                                        "type": "string"
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "reservation_status": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/clients/totalCostAbove/{price}": {
            "get": {
                "operationId": "totalCostAbove",
                "tags": [
                    "Clients"
                ],
                "description": "Recupere la liste des clients dont le cout total des reservations est superieur a un prix donne",
                "parameters": [
                    {
                        "name": "price",
                        "in": "path",
                        "required": true,
                        "type": "number",
                        "format": "float"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id_client": {
                                        "type": "integer"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "registration_date": {
                                        "type": "string"
                                    },
                                    "id_reservation": {
                                        "type": "integer"
                                    },
                                    "id_room": {
                                        "type": "integer"
                                    },
                                    "checkin_date": {
                                        "type": "string"
                                    },
                                    "checkout_date": {
                                        "type": "string"
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "reservation_status": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/clients/roomType/{roomType}": {
            "get": {
                "operationId": "clientsByRoomType",
                "tags": [
                    "Clients"
                ],
                "description": "Recupere la liste des clients ayant reserve une chambre d'un type donne",
                "parameters": [
                    {
                        "name": "roomType",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id_client": {
                                        "type": "integer"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "registration_date": {
                                        "type": "string"
                                    },
                                    "id_reservation": {
                                        "type": "integer"
                                    },
                                    "id_room": {
                                        "type": "integer"
                                    },
                                    "checkin_date": {
                                        "type": "string"
                                    },
                                    "checkout_date": {
                                        "type": "string"
                                    },
                                    "total_price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "reservation_status": {
                                        "type": "string"
                                    },
                                    "room_number": {
                                        "type": "integer"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    },
                                    "price_per_night": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    },
                                    "room_type": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/clients/maxAmountSpent": {
            "get": {
                "operationId": "maxAmountSpent",
                "tags": [
                    "Clients"
                ],
                "description": "Recupere le client ayant depense le plus",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "client": {
                                        "type": "string"
                                    },
                                    "total": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    }
}